Detection/requirements.txt
asone==0.2.8
click==8.1.3
Cython==0.29.34
cython-bbox==0.1.3
IPython==8.12.0
numpy==1.23.3
onnxruntime-gpu==1.12.1
opencv-python==4.7.0.72
Pillow==9.5.0
python-dotenv==1.0.0
ultralytics==8.0.61